Change to skipUnless, refactor to base class
