+++ test/connectors/test_cpu_torch_networks.py
@@ -15,15 +15,10 @@ import unittest
 from test import QiskitMachineLearningTestCase
 from test.connectors.test_torch_networks import TestTorchNetworks
 
-from qiskit.utils import optionals
-import qiskit_machine_learning.optionals as _optionals
-
 
 class TestCPUTorchNetworks(QiskitMachineLearningTestCase, TestTorchNetworks):
     """CPU based tests of hybrid PyTorch networks."""
 
-    @unittest.skipUnless(optionals.HAS_AER, "qiskit-aer is required to run this test")
-    @unittest.skipIf(not _optionals.HAS_TORCH, "PyTorch not available.")
     def setUp(self):
         super().setup_test()
         super().setUp()
