+++ test/connectors/test_gpu_torch_connector.py
@@ -15,15 +15,10 @@ import unittest
 from test import QiskitMachineLearningTestCase, gpu
 from test.connectors.test_torch_connector import TestTorchConnector
 
-from qiskit.utils import optionals
-import qiskit_machine_learning.optionals as _optionals
-
 
 class TestGPUTorchConnector(QiskitMachineLearningTestCase, TestTorchConnector):
     """GPU based tests of the PyTorch connector."""
 
-    @unittest.skipUnless(optionals.HAS_AER, "qiskit-aer is required to run this test")
-    @unittest.skipIf(not _optionals.HAS_TORCH, "PyTorch not available.")
     @gpu
     def setUp(self):
         super().setup_test()
